requirements.txt
flask
ffmpeg-python
pytesseract
pdfplumber
python-docx
pandas
transformers
torch
